task_continuity/requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
python-dateutil>=2.8.2
pydantic>=2.5.0
pydantic-settings>=2.0.0